# backend/docs/swagger.yaml
basePath: /api
definitions:
  branch.Branch:
    properties:
      capacity:
        example: 30
        type: integer
      id:
        example: 1
        type: integer
      name:
        example: Siam
        type: string
    type: object
  branch.UpdateCapacityRequest:
    properties:
      capacity:
        example: 30
        type: integer
    required:
    - capacity
    type: object
  commute.CommuteResponse:
    properties:
      commute_time:
        example: 30
        type: integer
      destination_branch:
        example: 2
        type: integer
      source_branch:
        example: 1
        type: integer
    type: object
  commute.UpdateCommuteRequest:
    properties:
      commute_time:
        example: 30
        type: integer
    required:
    - commute_time
    type: object
  scheduling.Booking:
    properties:
      branch_id:
        type: integer
      created_at:
        type: string
      end_time:
        type: string
      id:
        type: integer
      start_time:
        type: string
      student_id:
        type: integer
      student_name:
        type: string
      subject_id:
        type: integer
      teacher_id:
        type: integer
    type: object
...
  scheduling.BookingRequest:
    properties:
      branch_id:
        example: 1
        type: integer
      duration_minutes:
        example: 60
        type: integer
      preferred_slots:
        items:
          $ref: '#/definitions/shared.WeeklySlot'
        type: array
      preferred_teacher_id:
        example: 5
        type: integer
      required_gender:
        enum:
        - male
        - female
        - lgbtq+
        example: female
        type: string
      subject_id:
        example: 1
        type: integer
    required:
    - branch_id
    - preferred_slots
    - required_gender
    - subject_id
    type: object
  scheduling.BookingResponse:
    properties:
      results:
        items:
          $ref: '#/definitions/scheduling.SlotResult'
        type: array
    type: object
  scheduling.ConfirmBookingRequest:
    properties:
      branch_id:
        example: 1
        type: integer
      end_time:
        example: "2026-06-01T10:00:00Z"
        type: string
      start_time:
        example: "2026-06-01T09:00:00Z"
        type: string
      student_id:
        example: 3
        type: integer
      subject_id:
        example: 1
        type: integer
      teacher_id:
        example: 1
        type: integer
    required:
    - branch_id
    - end_time
    - start_time
    - student_id
    - subject_id
    - teacher_id
    type: object
  scheduling.ErrorResponse:
    properties:
      error:
        example: failed to retrieve teachers
        type: string
    type: object
  scheduling.MessageResponse:
    properties:
      message:
        example: Availability saved successfully
        type: string
    type: object
...
  shared.Subject:
    properties:
      id:
        type: integer
      name:
        type: string
    type: object
...
  teacher.TeacherAvailability:
    properties:
      teacher:
        $ref: '#/definitions/teacher.Teacher'
      weekly:
        items:
          $ref: '#/definitions/shared.WeeklySlot'
        type: array
    type: object
  teacher.UpdateGenderRequest:
    properties:
      gender:
        example: female
        type: string
    required:
    - gender
    type: object
...
  user.AuthResponse:
    properties:
      token:
        example: eyJhbGciOiJIUzI1NiIs...
        type: string
      user:
        $ref: '#/definitions/user.User'
    type: object
  user.LinkStudentRequest:
    properties:
      student_id:
        example: 3
        type: integer
    required:
    - student_id
    type: object
  user.LoginRequest:
    properties:
      name:
        example: alice
        type: string
      password:
        example: s3cret
        type: string
    required:
    - name
    - password
    type: object
  user.ParentWithStudents:
    properties:
      id:
        example: 1
        type: integer
      name:
        example: alice
        type: string
      role:
        example: student
        type: string
      students:
        items:
          $ref: '#/definitions/user.User'
        type: array
    type: object
...
  user.User:
    properties:
      id:
        example: 1
        type: integer
      name:
        example: alice
        type: string
      role:
        example: student
        type: string
    type: object
info:
  contact: {}
  description: API for teacher availability management and bookings
  title: Converge API
  version: "1.0"
paths:
  /availability:
    get:
      description: Returns all active teachers with their weekly availability slots,
        grouped by teacher
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/teacher.TeacherAvailability'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/scheduling.ErrorResponse'
      summary: Get all teacher availability
      tags:
      - availability
    post:
      consumes:
      - application/json
      description: Replaces all existing availability for a teacher with the provided
        weekly schedule (full replace, not merge). Rejects overlapping slots within
        the same day.
      parameters:
      - description: Weekly availability payload
        in: body
        name: body
        required: true
        schema:
          additionalProperties: true
          type: object
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/scheduling.MessageResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/scheduling.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/scheduling.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/scheduling.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/scheduling.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Submit weekly availability
      tags:
      - availability
  /bookings:
    get:
      description: Admin sees all bookings; a student sees only their own; a parent
        sees their students' bookings.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/scheduling.Booking'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/scheduling.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/scheduling.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/scheduling.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List bookings (scoped by role)
      tags:
      - bookings
    post:
      consumes:
      - application/json
      description: Evaluates each preferred_slot independently and returns per-slot
        results with either an exact_match or alternatives.
      parameters:
      - description: Booking request
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/scheduling.BookingRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/scheduling.BookingResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/scheduling.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/scheduling.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/scheduling.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/scheduling.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Evaluate a booking request
      tags:
      - bookings
  /bookings/{id}:
    delete:
      description: Deletes a confirmed booking by ID. Returns 404 if not found.
      parameters:
      - description: Booking ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/scheduling.MessageResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/scheduling.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/scheduling.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/scheduling.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/scheduling.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/scheduling.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Cancel a booking
      tags:
      - bookings
  /bookings/confirm:
    post:
      consumes:
      - application/json
      description: Creates the actual booking in the database after the client selects
        a result from the evaluate endpoint.
      parameters:
      - description: Booking confirmation
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/scheduling.ConfirmBookingRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/scheduling.Booking'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/scheduling.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/scheduling.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/scheduling.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/scheduling.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/scheduling.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Confirm a booking
      tags:
      - bookings
  /branches:
    get:
      description: Returns all branches ordered by id
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/branch.Branch'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/scheduling.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/scheduling.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/scheduling.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List branches
      tags:
      - branches
  /branches/{id}/capacity:
    patch:
      consumes:
      - application/json
      description: Sets how many bookings a branch can hold concurrently. 0 means
        unlimited/unenforced.
      parameters:
      - description: Branch ID
        in: path
        name: id
        required: true
        type: integer
      - description: New capacity
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/branch.UpdateCapacityRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/scheduling.MessageResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/scheduling.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/scheduling.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/scheduling.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/scheduling.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/scheduling.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update a branch's capacity
      tags:
      - branches
  /commute:
    get:
      description: Returns the commute time in minutes. Without query params, returns
        the default commute time. With both source_branch and destination_branch,
        returns 0 when they are equal, otherwise the default.
      parameters:
      - description: Source branch ID
        in: query
        name: source_branch
        type: integer
      - description: Destination branch ID
        in: query
        name: destination_branch
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/commute.CommuteResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/scheduling.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/scheduling.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/scheduling.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/scheduling.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get commute time between two branches
      tags:
      - commute
    patch:
      consumes:
      - application/json
      description: Sets the commute time in minutes applied between different branches.
        Must be non-negative.
      parameters:
      - description: New commute time
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/commute.UpdateCommuteRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/commute.CommuteResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/scheduling.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/scheduling.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/scheduling.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/scheduling.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update the global commute time
      tags:
      - commute
  /login:
    post:
      consumes:
      - application/json
      description: Verifies a user's name and password and returns a JWT access token
        plus the user.
      parameters:
      - description: User credentials
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/user.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/user.AuthResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/scheduling.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/scheduling.ErrorResponse'
      summary: Log in
      tags:
      - users
  /parents:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/user.ParentWithStudents'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/scheduling.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/scheduling.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/scheduling.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List all parents with their students (admin only)
      tags:
      - users
  /parents/{id}/students:
    get:
      parameters:
      - description: Parent user ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/user.User'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/scheduling.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/scheduling.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List a parent's linked students (admin only)
      tags:
      - users
    post:
      consumes:
      - application/json
      parameters:
      - description: Parent user ID
        in: path
        name: id
        required: true
        type: integer
      - description: Student to link
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/user.LinkStudentRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/scheduling.MessageResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/scheduling.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/scheduling.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/scheduling.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Link a student to a parent (admin only)
      tags:
      - users
  /parents/{id}/students/{studentId}:
    delete:
      parameters:
      - description: Parent user ID
        in: path
        name: id
        required: true
        type: integer
      - description: Student user ID
        in: path
        name: studentId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/scheduling.MessageResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/scheduling.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/scheduling.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/scheduling.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Unlink a student from a parent (admin only)
      tags:
      - users
  /register:
    post:
      consumes:
      - application/json
      description: Admin-only. Creates a user with a unique name, a bcrypt-hashed
        password, and the given role.
      parameters:
      - description: New user credentials and role
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/user.RegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/user.User'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/scheduling.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/scheduling.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/scheduling.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/scheduling.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/scheduling.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Register a new user (admin only)
      tags:
      - users
  /students:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/user.User'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/scheduling.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/scheduling.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/scheduling.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List all students (admin only)
      tags:
      - users
  /subjects:
    get:
      description: Returns all subjects ordered by id
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/shared.Subject'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/scheduling.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/scheduling.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/scheduling.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List subjects
      tags:
      - subjects
  /teachers:
    get:
      description: Returns teachers ordered by name. Filter by status (active|deactivated|all,
        default all). Optionally filter by subject_id, which returns active teachers
        only and ignores status.
      parameters:
      - description: Filter by subject ID (active teachers only)
        in: query
        name: subject_id
        type: integer
      - description: Filter by status
        enum:
        - active
        - deactivated
        - all
        in: query
        name: status
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/teacher.Teacher'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/scheduling.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/scheduling.ErrorResponse'
      summary: List teachers
      tags:
      - teachers
    post:
      consumes:
      - application/json
      description: Adds a new teacher with name, email, and gender
      parameters:
      - description: New teacher payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/teacher.CreateTeacherRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/teacher.Teacher'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/scheduling.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/scheduling.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/scheduling.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/scheduling.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create a teacher
      tags:
      - teachers
  /teachers/{id}/gender:
    patch:
      consumes:
      - application/json
      description: Sets a teacher's gender to male, female, or lgbtq+
      parameters:
      - description: Teacher ID
        in: path
        name: id
        required: true
        type: integer
      - description: New gender
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/teacher.UpdateGenderRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/scheduling.MessageResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/scheduling.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/scheduling.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/scheduling.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/scheduling.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/scheduling.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update a teacher's gender
      tags:
      - teachers
  /teachers/{id}/status:
    patch:
      consumes:
      - application/json
      description: Sets a teacher's status to active or deactivated
      parameters:
      - description: Teacher ID
        in: path
        name: id
        required: true
        type: integer
      - description: New status
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/teacher.UpdateStatusRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/scheduling.MessageResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/scheduling.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/scheduling.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/scheduling.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/scheduling.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/scheduling.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Toggle teacher active/deactivated status
      tags:
      - teachers
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and the JWT from /login.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"